requirements.txt
# 核心依赖
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0

# 缓存
cachetools>=5.3.0
redis>=5.0.0  # Redis缓存支持

# SSE支持
sse-starlette>=1.8.0  # 服务器发送事件支持

# 记忆系统
chromadb>=0.4.0
sentence-transformers>=2.2.0

# LLM 集成
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5

# 工具库
python-dotenv>=1.0.0
pyyaml>=6.0
aiofiles>=23.0.0
httpx>=0.25.0  # 异步HTTP客户端

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
faker>=20.0.0

# 系统监控
psutil>=5.9.0
